requirements.txt
Django
behave-django
factory-boy
selenium
